Remove trailing whitespace (#96)
I am being bold now. There are already many translations without this trailing whitespace so we may as well remove it from the source string. In case this appears to be wrong a follow up pull needs to be done anyways to create a situation where the whitespace is not needed anyways.

[skip ci]