Make extension work with MW 1.39 and clean up a bit
In addition to replacing uses of deprecated or removed functions/types,
notable changes include

- Addition of some logging to the (newly introduced) `SemanticWatchlist` log
group.

- Tweaking the JavaScript for Special:WatchlistConditions to avoid multiple
requests in parallel that could cause database problems.

- Switching the hook handling to handle the `SaveUserOptions` and
`SkinTemplateNavigationUniversal` rather than `UserSaveOptions` and
`PersonalUrls`, both of which were removed. Handlers were left with the old
class names and file names to allow git to track the history, and will be
renamed in a follow-up.

At this point, there still remains cleanup to be done, but the extension is
in a working form.