Additional compatibility fixes for MW 1.39, fix group coverage logic
Previously, a group with no namespaces set would cover every page, same with
groups with no categories or no concepts. Instead, when an option is not set
it now no longer indicates coverage.

Includes replacing uses of MWNamespace with the NamespaceInfo service, and
using the fully qualified name of some SemanticMediaWiki classes.