i18n: define missing `action-*` messages
Add messages `action-semanticwatch` and `action-semanticwatchgroups`, which are
used to display error messages on Special:SemanticWatchlist and
Special:WatchlistConditions respectively when the viewer does not have the
rights needed to access the pages.

Note that unlike `right-semanticwatchgroups`, the action message does not
include a link to the special page, since it is shown while on the special
page.

METHW-112